Cleanup VSCode docs and script
I'm always changing VSCode extensions and themes, so there's no need to
be keeping track of those and homepages. It's easy enough to get
information about an extension just by searching for it, so the
documentation file for VSCode is gone.

So are some less used extensions. And since I quite like Kotlin, and
sometines opening IntelliJ is overkill, Kotlin extension for VSCode
is helpful.

Closes #17